Create corrected unique genre count per year CSV
This is the most accurate data for unique genres in our streaming data. The previous CSV was subject to exceptions and issues connecting to the Spotify API, coupled with the random-access nature of obtaining query results, meaning it was effectively a random sample. This CSV was generated after reprocessing the entire `artists` table, and filling any gaps in data due to problems with our Spotify token or API rate-limiting.